Split the test to reduce CI time
