+++ .github/workflows/build-test.yml
@@ -32,7 +32,6 @@ jobs:
       matrix:
         os: [ubuntu-20.04]
         malloc: [dlmalloc, mimalloc]
-        # metadata: [etcd, redis]
         metadata: [etcd]  # disable redis for now as it seems has some nondeterministic bugs
         exclude:
           - malloc: mimalloc
@@ -42,7 +41,7 @@ jobs:
     steps:
       - uses: actions/checkout@v3
         with:
-          submodules: recursive
+          submodules: true
 
       - name: Generate Summary for Submodules
         run: |
@@ -140,19 +139,6 @@ jobs:
         if: false
         uses: mxschmitt/action-tmate@v3
 
-      - name: Install libgrape-lite
-        if: runner.os == 'Linux'
-        run: |
-          export PATH=/usr/lib/ccache:$PATH
-
-          git clone https://github.com/alibaba/libgrape-lite.git
-          cd libgrape-lite
-          mkdir build
-          cd build
-          cmake ..
-          make -j`nproc`
-          sudo make install
-
       - name: Setup SSH environment
         run: |
           # avoid access node by DNS hostnames
@@ -213,7 +199,8 @@ jobs:
                    -DBUILD_VINEYARD_HOSSEINMOEIN_DATAFRAME=ON \
                    -DBUILD_VINEYARD_FUSE=ON \
                    -DBUILD_VINEYARD_FUSE_PARQUET=ON \
-                   -DBUILD_VINEYARD_GRAPH_WITH_GAR=ON \
+                   -DBUILD_VINEYARD_GRAPH=OFF \
+                   -DBUILD_VINEYARD_GRAPH_WITH_GAR=OFF \
                    -DBUILD_VINEYARD_TESTS_ALL=ON
 
           if [ "${{ matrix.metadata }}" == "redis" ]; then
@@ -290,13 +277,6 @@ jobs:
 
       - name: Build
         run: |
-          # Workaround to [Errno 28] No space left on device
-          # https://github.com/actions/runner-images/issues/2840#issuecomment-790492173
-          sudo rm -rf /usr/share/dotnet
-          sudo rm -rf /opt/ghc
-          sudo rm -rf "/usr/local/share/boost"
-          sudo rm -rf "$AGENT_TOOLSDIRECTORY"
-
           export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib:/usr/local/lib64
 
           pushd build
@@ -323,7 +303,7 @@ jobs:
           twine check dist/*.whl
           sudo pip3 install dist/*.whl -U
 
-      - name: Get Test Data
+      - name: Get test data
         run: |
           git clone https://github.com/7br/gstest.git --depth=1
 
@@ -331,7 +311,7 @@ jobs:
         if: false
         uses: mxschmitt/action-tmate@v3
 
-      - name: Run Tests
+      - name: Run tests
         run: |
           export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib:/usr/local/lib64:/usr/local/lib/x86_64-linux-gnu
 
@@ -346,7 +326,7 @@ jobs:
         if: false
         uses: mxschmitt/action-tmate@v3
 
-      - name: Run Python Tests with Debug
+      - name: Run python tests with debug
         if: ${{ github.event_name == 'push' }}
         uses: sighingnow/action-tmate@master
         with:
@@ -369,7 +349,7 @@ jobs:
               python3 test/runner.py $RUNNER_ARGS --with-contrib
             fi
 
-      - name: Run Python Tests
+      - name: Run python tests
         if: ${{ github.event_name == 'pull_request' }}
         run: |
           export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib:/usr/local/lib64:/usr/local/lib/x86_64-linux-gnu
@@ -394,7 +374,7 @@ jobs:
         if: false
         uses: mxschmitt/action-tmate@v3
 
-      - name: Run IO Tests
+      - name: Run I/O tests
         run: |
           export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib:/usr/local/lib64:/usr/local/lib/x86_64-linux-gnu
 
@@ -407,7 +387,7 @@ jobs:
           rm -rf /dev/shm/etcd*
           python3 test/runner.py $RUNNER_ARGS --with-io --with-migration
 
-      - name: Run FUSE Tests
+      - name: Run FUSE tests
         run: |
           export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib:/usr/local/lib64:/usr/local/lib/x86_64-linux-gnu
