+++ modules/graph/CMakeLists.txt
@@ -129,6 +129,9 @@ add_executable(vineyard-graph-loader ${GRAPH_LOADER_SRC_FILES})
 target_link_libraries(vineyard-graph-loader PRIVATE vineyard_graph)
 install_vineyard_target(vineyard-graph-loader)
 
+add_custom_target(vineyard_graph_tests)
+add_dependencies(vineyard_tests vineyard_graph_tests)
+
 if(BUILD_VINEYARD_TESTS)
     enable_testing()
     file(GLOB TEST_FILES RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}/test" "${CMAKE_CURRENT_SOURCE_DIR}/test/*.cc")
@@ -149,6 +152,6 @@ if(BUILD_VINEYARD_TESTS)
             target_link_libraries(${T_NAME} PRIVATE ${LIBUNWIND_LIBRARIES})
         endif()
         add_test(${T_NAME}, ${T_NAME})
-        add_dependencies(vineyard_tests ${T_NAME})
+        add_dependencies(vineyard_graph_tests ${T_NAME})
     endforeach()
 endif()
