+++ setup_ml.py
@@ -85,12 +85,7 @@ setup(
     packages=find_ml_packages('python'),
     cmdclass={'bdist_wheel': bdist_wheel_plat, 'install': install_plat},
     zip_safe=False,
-    install_requires=['vineyard', 'tensorflow', 'xgboost', 'torch', 'mxnet'],
-    extras_require={
-        'cuda': [
-            'nvidia-dali-cuda110',
-        ],
-    },
+    install_requires=['vineyard'],
     platform=['POSIX', 'MacOS'],
     license="Apache License 2.0",
     classifiers=[
